refactor: separate student from school
