+++ src/main/kotlin/com/sexysisters/tojserverv2/domain/student/service/StudentService.kt
@@ -4,4 +4,6 @@ import com.sexysisters.tojserverv2.domain.school.SchoolCommand
 
 interface StudentService {
     fun createStudent(command: SchoolCommand.CreateStudent): Long
+    fun joinSchool(code: String, studentId: Long): String
+    fun applySchool(code: String, studentId: Long): String
 }
